requirements.txt
h5py==2.10.0
numpy==1.18.1
pandas==0.25.3
scikit-learn==0.22.1
scipy==1.4.1
tensorflow-gpu==2.0.1